Restructuring: added datasets/ and utils/ under torch_datasets/
